GUI ver. 3.1

+++ G_U_I/src/application/Autokary.java
@@ -5,7 +5,7 @@ public class Autokary {
 	private String AUT_Marka;
 	private String AUT_Model;
 	private String AUT_Rok_Prod;
-	private float AUT_Poj_silnik;
+	private Double AUT_Poj_silnik;
 	private int AUT_Spalamie_lkm;
 	private String AUT_Kat_autokaru;
 	private int AUT_Ilosc_miejsc;
@@ -17,14 +17,14 @@ public class Autokary {
 		this.AUT_Marka = "";
 		this.AUT_Model = "";
 		this.AUT_Rok_Prod = "";
-		this.AUT_Poj_silnik = 0;
+		this.AUT_Poj_silnik = 0.0;
 		this.AUT_Spalamie_lkm =0;
 		this.AUT_Kat_autokaru = "";
 		this.AUT_Ilosc_miejsc = 0;
 		this.AUT_Nr_rejestracji = "";
 	}
 	
-public  Autokary( int AUT_KEY, String AUT_Marka, String AUT_Model, String AUT_Rok_Prod, float AUT_Poj_silnik, int AUT_Spalamie_lkm, String AUT_Kat_autokaru, int AUT_Ilosc_miejsc, String AUT_Nr_rejestracji)
+public  Autokary( int AUT_KEY, String AUT_Marka, String AUT_Model, String AUT_Rok_Prod,  Double AUT_Poj_silnik, int AUT_Spalamie_lkm, String AUT_Kat_autokaru, int AUT_Ilosc_miejsc, String AUT_Nr_rejestracji)
 	{
  
 	this.AUT_KEY =AUT_KEY;
@@ -72,11 +72,11 @@ public void setAUT_Rok_Prod(String aUT_Rok_Prod) {
 	AUT_Rok_Prod = aUT_Rok_Prod;
 }
 
-public float getAUT_Poj_silnik() {
+public Double getAUT_Poj_silnik() {
 	return AUT_Poj_silnik;
 }
 
-public void setAUT_Poj_silnik(float aUT_Poj_silnik) {
+public void setAUT_Poj_silnik(Double aUT_Poj_silnik) {
 	AUT_Poj_silnik = aUT_Poj_silnik;
 }
